Clean code

+++ src/main/java/com/jhlabs/composite/BurnComposite.java
@@ -16,7 +16,7 @@ limitations under the License.
 package com.jhlabs.composite;
 
 import java.awt.*;
-import java.awt.image.*;
+import java.awt.image.ColorModel;
 
 public final class BurnComposite extends RGBComposite {
 
@@ -49,21 +49,18 @@ public final class BurnComposite extends RGBComposite {
                 int dor, dog, dob;
 
                 if (dir != 255) {
-                    dor = clamp(255 - (((int) (255 - sr) << 8) / (dir + 1)));
-                }
-                else {
+                    dor = clamp(255 - (((255 - sr) << 8) / (dir + 1)));
+                } else {
                     dor = sr;
                 }
                 if (dig != 255) {
-                    dog = clamp(255 - (((int) (255 - sg) << 8) / (dig + 1)));
-                }
-                else {
+                    dog = clamp(255 - (((255 - sg) << 8) / (dig + 1)));
+                } else {
                     dog = sg;
                 }
                 if (dib != 255) {
-                    dob = clamp(255 - (((int) (255 - sb) << 8) / (dib + 1)));
-                }
-                else {
+                    dob = clamp(255 - (((255 - sb) << 8) / (dib + 1)));
+                } else {
                     dob = sb;
                 }
