+++ src/main/java/com/jhlabs/composite/ColorBurnComposite.java
@@ -16,7 +16,7 @@ limitations under the License.
 package com.jhlabs.composite;
 
 import java.awt.*;
-import java.awt.image.*;
+import java.awt.image.ColorModel;
 
 public final class ColorBurnComposite extends RGBComposite {
 
@@ -49,21 +49,18 @@ public final class ColorBurnComposite extends RGBComposite {
                 int dor, dog, dob;
 
                 if (sr != 0) {
-                    dor = Math.max(255 - (((int) (255 - dir) << 8) / sr), 0);
-                }
-                else {
+                    dor = Math.max(255 - (((255 - dir) << 8) / sr), 0);
+                } else {
                     dor = sr;
                 }
                 if (sg != 0) {
-                    dog = Math.max(255 - (((int) (255 - dig) << 8) / sg), 0);
-                }
-                else {
+                    dog = Math.max(255 - (((255 - dig) << 8) / sg), 0);
+                } else {
                     dog = sg;
                 }
                 if (sb != 0) {
-                    dob = Math.max(255 - (((int) (255 - dib) << 8) / sb), 0);
-                }
-                else {
+                    dob = Math.max(255 - (((255 - dib) << 8) / sb), 0);
+                } else {
                     dob = sb;
                 }
