+++ src/main/java/com/jhlabs/image/EqualizeFilter.java
@@ -41,11 +41,10 @@ public class EqualizeFilter extends WholeImageFilter {
                     lut[i][j] = lut[i][j - 1] + histogram.getFrequency(i, j);
                 }
                 for (j = 0; j < 256; j++) {
-                    lut[i][j] = (int) Math.round(lut[i][j] * scale);
+                    lut[i][j] = Math.round(lut[i][j] * scale);
                 }
             }
-        }
-        else {
+        } else {
             lut = null;
         }
